requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.3
python-whois==0.8.0
tldextract==3.4.4
urllib3==2.0.7
certifi==2023.7.22
joblib==1.3.2 